Refactor remainder of table sections
